# packages/client/swagger/swagger.yaml
openapi: 3.0.0
paths:
...
  /boards/nodes:
    post:
      description: Nodes successfully created
      operationId: NodesController_createMany
      parameters:
        - description: Board id where nodes should be added
          name: boardId
          in: query
          required: true
          schema:
            format: uuid
            type: string
        - description: List of nodes to create
          name: nodes
          in: query
          required: true
          schema:
            type: array
            items:
              type: array
              items:
                type: nodeApiType
      responses:
        '201':
          description: ''
      security:
        - cookie: []
      status: 201
      summary: ''
      tags:
        - Nodes
    patch:
      description: Nodes successfully updated
      operationId: NodesController_updateMany
      parameters:
        - description: Board id where nodes should be updated
          name: boardId
          in: query
          required: true
          schema:
            format: uuid
            type: string
        - description: List of nodes to updated
          name: nodes
          in: query
          required: true
          schema:
            type: array
            items:
              type: array
              items:
                type: nodeApiType
      responses:
        '200':
          description: ''
      security:
        - cookie: []
      status: 200
      summary: ''
      tags:
        - Nodes
    delete:
      description: Nodes successfully removed
      operationId: NodesController_removeMany
      parameters:
        - description: Board id where nodes should be updated
          name: boardId
          in: query
          required: true
          schema:
            type: string
        - description: List of node ids to remove
          name: ids
          in: query
          required: true
          schema:
            type: array
            items:
              type: array
              items:
                type: string
      responses:
        '204':
          description: ''
      security:
        - cookie: []
      status: 204
      summary: ''
      tags:
        - Nodes
  /boards/nodes/{boardId}:
    get:
      description: Nodes successfully found
      isArray: true
      operationId: NodesController_findAll
      parameters:
        - name: boardId
          required: true
          in: path
          description: Board id where nodes are located
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: ''
      security:
        - cookie: []
      status: 200
      summary: ''
      tags:
        - Nodes
      type: []
...
info:
  title: Gateway
  description: Gateway Service
  version: 1.0.0
  contact: {}
...
servers: []
...